Refactor ollama client structure and update exports
