+++ src/common/ollama/index.ts
@@ -1,0 +1,1 @@
+export {ollama, ollamaModel} from './ollamaClient';
